refactor doc_en
