--- docs/_static/css/readthedocs.css
@@ -1,6 +1,0 @@
-.header-logo {
-    background-image: url("../images/mmaction2.png");
-    background-size: 130px 40px;
-    height: 40px;
-    width: 130px;
-}
