+++ docs/en/_static/css/readthedocs.css
@@ -1,0 +1,6 @@
+.header-logo {
+    background-image: url("../images/logo.png");
+    background-size: 130px 40px;
+    height: 40px;
+    width: 130px;
+}
