+++ docs/en/stat.py
