+++ docs/en/user_guides/index.rst
@@ -1,0 +1,9 @@
+.. toctree::
+   :maxdepth: 1
+
+   config.md
+   inference.md
+   train_test.md.md
+   dataset_prepare.md
+   useful_tools.md
+   visualization.md
