--- docs/index.rst
@@ -1,75 +1,0 @@
-Welcome to MMAction2's documentation!
-=====================================
-
-You can switch between Chinese and English documents in the lower-left corner of the layout.
-
-您可以在页面左下角切换文档语言。
-
-.. toctree::
-   :maxdepth: 2
-
-   install.md
-   getting_started.md
-   demo.md
-   benchmark.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Datasets
-
-   datasets.md
-   data_preparation.md
-   supported_datasets.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Model Zoo
-
-   modelzoo.md
-   recognition_models.md
-   localization_models.md
-   detection_models.md
-   skeleton_models.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Tutorials
-
-   tutorials/1_config.md
-   tutorials/2_finetune.md
-   tutorials/3_new_dataset.md
-   tutorials/4_data_pipeline.md
-   tutorials/5_new_modules.md
-   tutorials/6_export_model.md
-   tutorials/7_customize_runtime.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Useful Tools and Scripts
-
-   useful_tools.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Notes
-
-   changelog.md
-   faq.md
-
-.. toctree::
-   :caption: API Reference
-
-   api.rst
-
-.. toctree::
-   :caption: Switch Language
-
-   switch_language.md
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`search`
