--- docs/projects.md
@@ -1,23 +1,0 @@
-# Projects based on MMAction2
-
-There are many research works and projects built on MMAction2.
-We list some of them as examples of how to extend MMAction2 for your own projects.
-As the page might not be completed, please feel free to create a PR to update this page.
-
-## Projects as an extension
-
-- [OTEAction2](https://github.com/openvinotoolkit/mmaction2): OpenVINO Training Extensions for Action Recognition.
-
-## Projects of papers
-
-There are also projects released with papers.
-Some of the papers are published in top-tier conferences (CVPR, ICCV, and ECCV), the others are also highly influential.
-To make this list also a reference for the community to develop and compare new video understanding algorithms, we list them following the time order of top-tier conferences.
-Methods already supported and maintained by MMAction2 are not listed.
-
-- Evidential Deep Learning for Open Set Action Recognition, ICCV 2021 Oral. [\[paper\]](https://arxiv.org/abs/2107.10161)[\[github\]](https://github.com/Cogito2012/DEAR)
-- Rethinking Self-supervised Correspondence Learning: A Video Frame-level Similarity Perspective, ICCV 2021 Oral. [\[paper\]](https://arxiv.org/abs/2103.17263)[\[github\]](https://github.com/xvjiarui/VFS)
-- MGSampler: An Explainable Sampling Strategy for Video Action Recognition, ICCV 2021. [\[paper\]](https://arxiv.org/abs/2104.09952)[\[github\]](https://github.com/MCG-NJU/MGSampler)
-- MultiSports: A Multi-Person Video Dataset of Spatio-Temporally Localized Sports Actions, ICCV 2021. [\[paper\]](https://arxiv.org/abs/2105.07404)
-- Video Swin Transformer. [\[paper\]](https://arxiv.org/abs/2106.13230)[\[github\]](https://github.com/SwinTransformer/Video-Swin-Transformer)
-- Long Short-Term Transformer for Online Action Detection. [\[paper\]](https://arxiv.org/abs/2107.03377)
