--- docs_zh_CN/README.md
@@ -1,1 +1,0 @@
-../README_zh-CN.md
