--- docs_zh_CN/api.rst
@@ -1,101 +1,0 @@
-mmaction.apis
--------------
-.. automodule:: mmaction.apis
-    :members:
-
-mmaction.core
--------------
-
-optimizer
-^^^^^^^^^
-.. automodule:: mmaction.core.optimizer
-    :members:
-
-evaluation
-^^^^^^^^^^
-.. automodule:: mmaction.core.evaluation
-    :members:
-
-scheduler
-^^
-.. automodule:: mmaction.core.scheduler
-    :members:
-
-mmaction.localization
----------------------
-
-localization
-^^^^^^^^^^^^
-.. automodule:: mmaction.localization
-    :members:
-
-mmaction.models
----------------
-
-models
-^^^^^^
-.. automodule:: mmaction.models
-    :members:
-
-recognizers
-^^^^^^^^^^^
-.. automodule:: mmaction.models.recognizers
-    :members:
-
-localizers
-^^^^^^^^^^
-.. automodule:: mmaction.models.localizers
-    :members:
-
-common
-^^^^^^
-.. automodule:: mmaction.models.common
-    :members:
-
-backbones
-^^^^^^^^^
-.. automodule:: mmaction.models.backbones
-    :members:
-
-heads
-^^^^^
-.. automodule:: mmaction.models.heads
-    :members:
-
-necks
-^^^^^
-.. automodule:: mmaction.models.necks
-    :members:
-
-losses
-^^^^^^
-.. automodule:: mmaction.models.losses
-    :members:
-
-mmaction.datasets
------------------
-
-datasets
-^^^^^^^^
-.. automodule:: mmaction.datasets
-    :members:
-
-pipelines
-^^^^^^^^^
-.. automodule:: mmaction.datasets.pipelines
-    :members:
-
-samplers
-^^^^^^^^
-.. automodule:: mmaction.datasets.samplers
-    :members:
-
-mmaction.utils
---------------
-.. automodule:: mmaction.utils
-    :members:
-
-mmaction.localization
----------------------
-.. automodule:: mmaction.localization
-    :members:
