--- docs_zh_CN/index.rst
@@ -1,74 +1,0 @@
-欢迎来到 MMAction2 的中文文档！
-=====================================
-
-您可以在页面左下角切换中英文文档。
-
-You can change the documentation language at the lower-left corner of the page.
-
-.. toctree::
-   :maxdepth: 2
-
-   install.md
-   getting_started.md
-   demo.md
-   benchmark.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: 数据集
-
-   datasets.md
-   data_preparation.md
-   supported_datasets.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: 模型库
-
-   modelzoo.md
-   recognition_models.md
-   localization_models.md
-   detection_models.md
-   skeleton_models.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: 教程
-
-   tutorials/1_config.md
-   tutorials/2_finetune.md
-   tutorials/3_new_dataset.md
-   tutorials/4_data_pipeline.md
-   tutorials/5_new_modules.md
-   tutorials/6_export_model.md
-   tutorials/7_customize_runtime.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: 实用工具和脚本
-
-   useful_tools.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: 记录
-
-   changelog.md
-   faq.md
-
-.. toctree::
-   :caption: API 参考文档
-
-   api.rst
-
-.. toctree::
-   :caption: 语言切换
-
-   switch_language.md
-
-
-索引和表格
-==================
-
-* :ref:`genindex`
-* :ref:`search`
