--- docs_zh_CN/switch_language.md
@@ -1,3 +1,0 @@
-## <a href='https://mmaction2.readthedocs.io/en/latest/'>English</a>
-
-## <a href='https://mmaction2.readthedocs.io/zh_CN/latest/'>简体中文</a>
